prepare for sysiamge excision
