+++ src/Pkg.jl
@@ -789,98 +789,6 @@ function _auto_precompile(ctx::Types.Context, pkgs::Vector{PackageSpec}=PackageS
     end
 end
 
-using LibGit2: LibGit2
-using Tar: Tar
-function _run_precompilation_script_setup()
-    tmp = mktempdir()
-    cd(tmp)
-    empty!(DEPOT_PATH)
-    pushfirst!(DEPOT_PATH, tmp)
-    touch("Project.toml")
-    Pkg.activate(".")
-    Pkg.generate("TestPkg")
-    uuid = TOML.parsefile(joinpath("TestPkg", "Project.toml"))["uuid"]
-    mv("TestPkg", "TestPkg.jl")
-    tree_hash = cd("TestPkg.jl") do
-        sig = LibGit2.Signature("TEST", "TEST@TEST.COM", round(time()), 0)
-        repo = LibGit2.init(".")
-        LibGit2.add!(repo, "")
-        commit = LibGit2.commit(repo, "initial commit"; author=sig, committer=sig)
-        th = LibGit2.peel(LibGit2.GitTree, LibGit2.GitObject(repo, commit)) |> LibGit2.GitHash |> string
-        close(repo)
-        th
-    end
-    # Prevent cloning the General registry by adding a fake one
-    mkpath("registries/Registry/T/TestPkg")
-    write("registries/Registry/Registry.toml", """
-        name = "Registry"
-        uuid = "37c07fec-e54c-4851-934c-2e3885e4053e"
-        repo = "https://github.com/JuliaRegistries/Registry.git"
-        [packages]
-        $uuid = { name = "TestPkg", path = "T/TestPkg" }
-        """)
-    write("registries/Registry/T/TestPkg/Compat.toml", """
-          ["0"]
-          julia = "1"
-          """)
-    write("registries/Registry/T/TestPkg/Deps.toml", """
-          ["0"]
-          Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
-          """)
-    write("registries/Registry/T/TestPkg/Versions.toml", """
-          ["0.1.0"]
-          git-tree-sha1 = "$tree_hash"
-          """)
-    write("registries/Registry/T/TestPkg/Package.toml", """
-        name = "TestPkg"
-        uuid = "$uuid"
-        repo = "$(escape_string(tmp))/TestPkg.jl"
-        """)
-    Tar.create("registries/Registry", "registries/Registry.tar")
-    cmd = `$(Pkg.PlatformEngines.exe7z()) a "registries/Registry.tar.gz" -tgzip "registries/Registry.tar"`
-    run(pipeline(cmd, stdout = stdout_f(), stderr = stderr_f()))
-    write("registries/Registry.toml", """
-          git-tree-sha1 = "11b5fad51c4f98cfe0c145ceab0b8fb63fed6f81"
-          uuid = "37c07fec-e54c-4851-934c-2e3885e4053e"
-          path = "Registry.tar.gz"
-    """)
-    Base.rm("registries/Registry"; recursive=true)
-    return tmp
-end
-
-function _run_precompilation_script_artifact()
-    # Create simple artifact, bind it, then use it:
-    foo_hash = Pkg.Artifacts.create_artifact(dir -> touch(joinpath(dir, "foo")))
-    Artifacts.bind_artifact!("./Artifacts.toml", "foo", foo_hash)
-    # Also create multiple platform-specific ones because that's a codepath we need precompiled
-    Artifacts.bind_artifact!("./Artifacts.toml", "foo_plat", foo_hash; platform=Base.BinaryPlatforms.HostPlatform())
-    # Because @artifact_str doesn't work at REPL-level, we JIT out a file that we can include()
-    write("load_artifact.jl", """
-          Pkg.Artifacts.artifact"foo"
-          Pkg.Artifacts.artifact"foo_plat"
-          """)
-    foo_path = include("load_artifact.jl")
-end
-
-const CTRL_C = '\x03'
-const precompile_script = """
-    import Pkg
-    _pwd = pwd()
-    Pkg.UPDATED_REGISTRY_THIS_SESSION[] = true
-    tmp = Pkg._run_precompilation_script_setup()
-    $CTRL_C
-    Pkg.add("TestPkg")
-    Pkg.develop(Pkg.PackageSpec(path="TestPkg.jl"))
-    Pkg.add(Pkg.PackageSpec(path="TestPkg.jl/"))
-    Pkg.REPLMode.try_prompt_pkg_add(Symbol[:notapackage])
-    Pkg.update(; update_registry=false)
-    Pkg.precompile()
-    ] add Te\t\t$CTRL_C
-    ] st
-    $CTRL_C
-    Pkg._run_precompilation_script_artifact()
-    rm(tmp; recursive=true)
-    cd(_pwd)
-    """
+include("precompile.jl")
 
 end # module
